--- src/main/java/org/c4marathon/assignment/domain/account/service/SettlementService.java
@@ -1,2 +1,0 @@
-package org.c4marathon.assignment.domain.account.service;public class SettlementService {
-}
