+++ src/main/java/org/c4marathon/assignment/domain/settlement/dto/response/SettlementResponseDto.java
@@ -1,7 +1,7 @@
-package org.c4marathon.assignment.domain.account.dto.response;
+package org.c4marathon.assignment.domain.settlement.dto.response;
 
-import org.c4marathon.assignment.domain.account.entity.settlement.SettlementStatus;
-import org.c4marathon.assignment.domain.account.entity.settlement.SettlementType;
+import org.c4marathon.assignment.domain.settlement.entity.settlement.SettlementStatus;
+import org.c4marathon.assignment.domain.settlement.entity.settlement.SettlementType;
 
 public class SettlementResponseDto {
 	private Long settlementId;
