+++ src/main/java/org/c4marathon/assignment/domain/settlement/repository/SettlementRepository.java
@@ -1,6 +1,6 @@
-package org.c4marathon.assignment.domain.account.repository;
+package org.c4marathon.assignment.domain.settlement.repository;
 
-import org.c4marathon.assignment.domain.account.entity.settlement.Settlement;
+import org.c4marathon.assignment.domain.settlement.entity.settlement.Settlement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SettlementRepository extends JpaRepository<Settlement, Long> {
